Move controlled_spring_mass_system to test_utilities
