+++ drake/systems/analysis/BUILD.bazel
@@ -106,7 +106,6 @@ drake_cc_googletest(
         ":runge_kutta3_integrator",
         ":simulator",
         "//drake/common/test_utilities:is_dynamic_castable",
-        "//drake/systems/analysis/test/controlled_spring_mass_system",
         "//drake/systems/analysis/test_utilities",
     ],
 )
