+++ drake/systems/analysis/test/simulator_test.cc
@@ -16,7 +16,7 @@
 #include "drake/systems/analysis/implicit_euler_integrator.h"
 #include "drake/systems/analysis/runge_kutta2_integrator.h"
 #include "drake/systems/analysis/runge_kutta3_integrator.h"
-#include "drake/systems/analysis/test/controlled_spring_mass_system/controlled_spring_mass_system.h"
+#include "drake/systems/analysis/test_utilities/controlled_spring_mass_system.h"
 #include "drake/systems/analysis/test_utilities/logistic_system.h"
 #include "drake/systems/analysis/test_utilities/my_spring_mass_system.h"
 #include "drake/systems/analysis/test_utilities/stateless_system.h"
