+++ drake/systems/analysis/test_utilities/controlled_spring_mass_system.h
